feat(topics): basic image cropping ui works
